usequery set

+++ src/Pages/Dashboard/AllFile.jsx
@@ -1,64 +1,68 @@
 import React, { useEffect, useState } from "react";
 import { BsPencilSquare, BsTrash } from "react-icons/bs";
+import { useQuery } from "react-query";
 import Swal from "sweetalert2";
+import Loading from "../Shared/Loading";
 import AddFile from "./AddFile";
 
 
 const AllFile = () => {
-  const [allFiles,setAllFiles] = useState([])
-  useEffect(()=>{
-    const url = `http://localhost:5000/files`
-    fetch(url, {
-        method: 'GET',
-        headers:{
-          "content-type": "application/json",
-        }
-    })
-    .then(res => res.json())
-    .then(data => {
-      setAllFiles(data)
-      
-    })
-      
-  },[])
+
+  const { isLoading, error, data:allFiles,refetch } = useQuery('allFiles', () =>
+    fetch(('http://localhost:5000/files'), {
+      method: 'GET',
+      headers: {
+        "content-type": "application/json",
+      }
+    }).then(res =>
+      res.json()
+    )    
+  )
+  refetch()
+
+  if (isLoading) {
+    return <Loading/>
+  }
+
+  if (error) return 'An error has occurred: ' + error.message
+  
   const handleDelete = id => {
     Swal.fire({
-        title: 'Are you sure?',
-        text: "You won't be able to revert this!",
-        icon: 'warning',
-        showCancelButton: true,
-        confirmButtonColor: '#3085d6',
-        cancelButtonColor: '#d33',
-        confirmButtonText: 'Yes, delete it!'
-      }).then((result) => {
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
 
-    if (result.value) {
+      if (result.value) {
         const url = `http://localhost:5000/files/${id}`;
         fetch(url, {
-            method: 'DELETE'
+          method: 'DELETE'
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data)
-                const remaining = allFiles.filter(product => product._id !== id);
-                setAllFiles(remaining)
-            })
+          .then(res => res.json())
+          .then(data => {
+            console.log(data)
+            const remaining = allFiles.filter(file => file._id !== id);
+          })
 
 
-    }
-    if (result.isConfirmed) {
+      }
+      if (result.isConfirmed) {
         Swal.fire(
           'Deleted!',
           'Your file has been deleted.',
           'success'
         )
       }
-})
-}
+    })
+  }
   return (
     <div>
       <div className="flex justify-end mb-5">
-        <AddFile/>
+        <AddFile />
       </div>
       <div className="overflow-x-auto">
         <table className="table table-zebra w-full">
@@ -76,23 +80,23 @@ const AllFile = () => {
           <tbody>
             {/* <!-- row 1 --> */}
             {
-              allFiles.map((file,index) => 
+              allFiles.map((file, index) =>
                 <tr>
-              <th>{index+1}</th>
-              <td>{file.title}</td>
-              <td>{file.model}</td>
-              <td>{file.year}</td>
-              <td>Pending</td>
-              <td className="flex gap-1">
-                <span>
-                  <BsPencilSquare />
-                </span>
-                <span>
-                  <BsTrash onClick={() => handleDelete(file._id)}/>
-                </span>
-              </td>
-            </tr>
-                )
+                  <th>{index + 1}</th>
+                  <td>{file.title}</td>
+                  <td>{file.model}</td>
+                  <td>{file.year}</td>
+                  <td>Pending</td>
+                  <td className="flex gap-1">
+                    <span>
+                      <BsPencilSquare />
+                    </span>
+                    <span>
+                      <BsTrash onClick={() => handleDelete(file._id)} />
+                    </span>
+                  </td>
+                </tr>
+              )
             }
             {/* <!-- row 2 --> */}
           </tbody>
